Switched to an sqllitedb for now.
I think this will just be easier than writing stuff pretending it will
eventually be talking a db. The way its written now you won't be able to
write any data to the db in a persistant when working through docker,
but thats easy enough to change later. So I'm going to hold off, since
we'll probably wnat to move the db to a different container anyways.